+++ src/jsonyx/_encoder.py
@@ -6,7 +6,6 @@ __all__: list[str] = ["Encoder"]
 import re
 import sys
 from io import StringIO
-from pathlib import Path
 from re import DOTALL, MULTILINE, VERBOSE, Match, Pattern, RegexFlag
 from typing import TYPE_CHECKING, Any, Protocol, TypeVar
 
@@ -412,38 +411,6 @@ class Encoder:
         )
         self._errors: str = "surrogatepass" if allow_surrogates else "strict"
 
-    def write(
-        self, obj: object, filename: _StrPath, encoding: str = "utf-8",
-    ) -> None:
-        r"""Serialize a Python object to a JSON file.
-
-        .. versionchanged:: 2.0 Added ``encoding``.
-
-        :param obj: a Python object
-        :param filename: the path to the JSON file
-        :param encoding: the JSON encoding
-        :raises OSError: if the file can't be opened
-        :raises RecursionError: if the object is too deeply nested
-        :raises TypeError: for unserializable values
-        :raises TruncatedSyntaxError: when failing to encode the file
-        :raises ValueError: for invalid values
-
-        Example:
-            >>> import jsonyx as json
-            >>> from pathlib import Path
-            >>> from tempfile import TemporaryDirectory
-            >>> encoder = json.Encoder()
-            >>> with TemporaryDirectory() as tmpdir:
-            ...     filename = Path(tmpdir) / "file.json"
-            ...     encoder.write(["filesystem API"], filename)
-            ...     filename.read_text("utf-8")
-            ...
-            '["filesystem API"]\n'
-
-        """
-        with Path(filename).open("w", -1, encoding, self._errors) as fp:
-            self.dump(obj, fp)
-
     def dump(self, obj: object, fp: _SupportsWrite[str] | None = None) -> None:
         r"""Serialize a Python object to an open JSON file.
 
@@ -471,9 +438,25 @@ class Encoder:
 
             >>> import jsonyx as json
             >>> from io import StringIO
+            >>> encoder = json.Encoder()
             >>> io = StringIO()
             >>> encoder.dump(["streaming API"], io)
             >>> io.getvalue()
+            '["streaming API"]\n'
+
+            Writing to a file:
+
+            >>> import jsonyx as json
+            >>> from os.path import join
+            >>> from tempfile import TemporaryDirectory
+            >>> encoder = json.Encoder()
+            >>> with TemporaryDirectory() as tmpdir:
+            ...     filename = join(tmpdir, "file.json")
+            ...     with open(filename, "w", encoding="utf-8") as fp:
+            ...         encoder.dump(["streaming API"], fp)
+            ...     with open(filename, "r", encoding="utf-8") as fp:
+            ...         fp.read()
+            ...
             '["streaming API"]\n'
 
         """